Implements GET, PUT, and HEAD for Objects.
There are two outstanding issues: if the content-type is set to "" (i.e.
the header is set but its value is ""), the PUT handler is not invoked
and 400 is returned. Filed a JIRA bug:
https://java.net/jira/browse/JERSEY-2823

HEAD for an object is also broken: Content-Length and Content-Type are
stripped by Jersey/Grizzly when processing the response. Filed a JIRA
bug: https://java.net/jira/browse/JERSEY-2822.

Tested with swift command line client. Will add unit tests in a
subsequent commit.